src/app/requirements.txt
flask
openai
sounddevice
scipy
numpy
pyserial
pyaudio
openwakeword
